+++ test.js
@@ -35,7 +35,7 @@ describe('jquery-app', function(done) {
         done();
     }));
 
-    it('should call plugin', createTestEnv(function (done, $) {
+    it('should call multiple plugins separated by comma', createTestEnv(function (done, $) {
         var foo = 0;
         var bar = 0;
 
@@ -43,6 +43,20 @@ describe('jquery-app', function(done) {
         $.fn.bar = function () { bar++; };
 
         $('body').append('<div data-plugin="foo, bar"></div>').app();
+
+        assert.equal(foo, 1);
+        assert.equal(bar, 1);
+        done();
+    }));
+
+    it('should call multiple plugins separated by whitespace', createTestEnv(function (done, $) {
+        var foo = 0;
+        var bar = 0;
+
+        $.fn.foo = function () { foo++; };
+        $.fn.bar = function () { bar++; };
+
+        $('body').append('<div data-plugin="foo bar"></div>').app();
 
         assert.equal(foo, 1);
         assert.equal(bar, 1);
@@ -124,6 +138,122 @@ describe('jquery-app', function(done) {
 
         assert.equal(foo, true);
         assert.equal(bar, true);
+        done();
+    }));
+
+    it('should call plugin with only specific namespace', createTestEnv(function (done, $) {
+        var foo = 0;
+        var bar = 0;
+
+        $.fn.foo = function () { foo++; };
+        $.fn.bar = function () { bar++; };
+
+        $('body').append('<div data-plugin="foo"><span data-other-plugin="bar"></span></div>').app({
+            'namespace': 'other-plugin'
+        });
+
+        assert.equal(foo, 0);
+        assert.equal(bar, 1);
+        done();
+    }));
+
+    it('should call plugins with only specific namespaces after dynamic DOM insert', createTestEnv(function (done, $) {
+        var foo = 0;
+        var bar = 0;
+
+        $.fn.foo = function () { foo++; };
+        $.fn.bar = function () { bar++; };
+
+        $('body').append('<div data-plugin="foo"><span data-other-plugin="bar"></span></div>');
+        $('body').app(); // data-plugin
+        $('body').app({'namespace': 'other-plugin'}); // data-other-plugin
+
+        $('body').append('<div data-plugin="foo"></div>');
+        $('body').app(); // data-plugin
+
+        assert.equal(foo, 2);
+        assert.equal(bar, 1);
+        done();
+    }));
+
+    it('$.app.hasPluginDefined returns true if plugin is defined', createTestEnv(function (done, $) {
+        $('body').append('<div data-plugin="foo"><span data-plugin="bar"></span></div>');
+        var foo = $('body').find('div[data-plugin]');
+        var bar = $('body').find('span[data-plugin]');
+
+        $.fn.foo = function () {};
+        $.fn.bar = function () {};
+
+        assert.equal($.app.hasPluginDefined(foo, 'foo'), true);
+        assert.equal($.app.hasPluginDefined(bar, 'foo'), false);
+        assert.equal($.app.hasPluginDefined(foo, 'bar'), false);
+        assert.equal($.app.hasPluginDefined(bar, 'bar'), true);
+
+        done();
+    }));
+
+    it('$.app.hasPluginDefined returns false if plugin doesn\'t exist', createTestEnv(function (done, $) {
+        $('body').append('<div data-plugin="foo"><span data-plugin="bar"></span></div>');
+        var foo = $('body').find('div[data-plugin]');
+        var bar = $('body').find('span[data-plugin]');
+
+        assert.equal($.app.hasPluginDefined(foo, 'foo'), false);
+        assert.equal($.app.hasPluginDefined(bar, 'bar'), false);
+
+        done();
+    }));
+
+    it('$.app.hasPlugin returns true if plugin is initialized', createTestEnv(function (done, $) {
+        $.fn.foo = function () {};
+        $.fn.bar = function () {};
+
+        $('body').append('<div data-plugin="foo baz"><span data-plugin="bar"></span></div>').app();
+        var foo = $('body').find('div[data-plugin]');
+        var bar = $('body').find('span[data-plugin]');
+
+        assert.equal($.app.hasPlugin(foo, 'baz'), false);
+        assert.equal($.app.hasPlugin(foo, 'foo'), true);
+        assert.equal($.app.hasPlugin(bar, 'foo'), false);
+        assert.equal($.app.hasPlugin(foo, 'bar'), false);
+        assert.equal($.app.hasPlugin(bar, 'bar'), true);
+
+        done();
+    }));
+
+    it('$.app.hasPlugin returns false if plugin doesn\'t exist', createTestEnv(function (done, $) {
+        $('body').append('<div data-plugin="foo baz"></div>').app();
+        var baz = $('body').find('div[data-plugin]');
+
+        assert.equal($.app.hasPlugin(baz, 'baz'), false);
+
+        done();
+    }));
+
+    it('$.app.getPlugins returns list of plugin names', createTestEnv(function (done, $) {
+        $.fn.foo = function () {};
+        $.fn.bar = function () {};
+        $.fn.baz = function () {};
+        $.fn.biz = function () {};
+        $.fn.bus = function () {};
+
+        $('body').append('<div data-plugin="foo, bar baz,biz, bus"></div>');
+        var element = $('body').find('div[data-plugin]');
+
+        assert.equal($.app.getPlugins(element).join(','), ['foo', 'bar', 'baz', 'biz', 'bus'].join(','));
+
+        done();
+    }));
+
+    it('$.app.getPlugins doesn\'t returns non-existing plugin names', createTestEnv(function (done, $) {
+        $.fn.foo = function () {};
+        $.fn.biz = function () {};
+        $.fn.bus = function () {};
+
+        $('body').append('<div data-plugin="foo, bar baz,biz, bus"></div>');
+        var element = $('body').find('div[data-plugin]');
+
+        assert.equal($.app.getPlugins(element).join(','), ['foo', 'biz', 'bus'].join(','));
+
         done();
     }));